Perform housekeeping, move tests to correct namespace
